permissions: refactor experiments permissions
