+++ cap/config.py
@@ -17,6 +17,15 @@ from os.path import dirname, join
 
 from celery.schedules import crontab
 from flask import request
+
+from cap.modules.deposit.permissions import (AdminDepositPermission,
+                                             CreateDepositPermission,
+                                             ReadDepositPermission)
+from cap.modules.oauthclient.contrib.cern import disconnect_handler
+from cap.modules.oauthclient.rest_handlers import (authorized_signup_handler,
+                                                   signup_handler)
+from cap.modules.records.permissions import ReadRecordPermission
+from cap.modules.search.facets import nested_filter
 from flask_principal import RoleNeed
 from invenio_deposit import config as deposit_config
 from invenio_deposit.config import DEPOSIT_REST_SORT_OPTIONS
@@ -30,15 +39,6 @@ from invenio_records_rest.facets import terms_filter
 from invenio_records_rest.utils import allow_all, deny_all
 from jsonresolver import JSONResolver
 from jsonresolver.contrib.jsonref import json_loader_factory
-
-from cap.modules.deposit.permissions import (AdminDepositPermission,
-                                             CreateDepositPermission,
-                                             ReadDepositPermission)
-from cap.modules.oauthclient.contrib.cern import disconnect_handler
-from cap.modules.oauthclient.rest_handlers import (authorized_signup_handler,
-                                                   signup_handler)
-from cap.modules.records.permissions import ReadRecordPermission
-from cap.modules.search.facets import nested_filter
 
 
 def _(x):
@@ -225,7 +225,7 @@ SUPERUSER_EGROUPS = [
 # CAP collaboration groups
 # ========================
 #: Configuration for collaborations
-CAP_COLLAB_EGROUPS = {
+EXPERIMENT_NEEDS = {
     "CMS": [
         RoleNeed("cms-members@cern.ch"),
     ],
@@ -237,15 +237,7 @@ CAP_COLLAB_EGROUPS = {
     ],
     "LHCb": [
         RoleNeed("lhcb-general@cern.ch"),
-    ],
-    "ADMIN": SUPERUSER_EGROUPS
-}
-
-EXPERIMENT_PERMISSION = {
-    "CMS": "cap.modules.experiments.permissions.cms.cms_group_need",
-    "ALICE": "cap.modules.experiments.permissions.alice.alice_group_need",
-    "ATLAS": "cap.modules.experiments.permissions.atlas.atlas_group_need",
-    "LHCb": "cap.modules.experiments.permissions.lhcb.lhcb_group_need",
+    ]
 }
 
 # Records
